MO : Improve ps_legalcompliance.php
From a practical point of view it makes no sense just to make the inclusion of the  _revocation terms_ or _privacy_ in the _terms of trade_ comfirmation dependent from the fact, if there is  any text content or not to be found in the cms "revocation terms" or "privacy". This is a conditio sine qua non, but *more important* is the fact that you click the option *Include revocation in terms and conditions*. 

Moreover, a confirmation of privacy in the checkout is only mandatory for Austria! Therefor I added the option for including a privacy statement to the _ps_legalcompliance_ configuration.